voglio tutti i prezzi

+++ src/main/java/it/prova/gestioneordini/dao/articolo/ArticoloDAO.java
@@ -8,6 +8,6 @@ public interface ArticoloDAO extends IBaseDAO<Articolo> {
 
 	public boolean verificaSeArticoloHaCategorie(long id) throws Exception;
 	
-	public int voglioSommaPrezziDiArticoliInQuellaCategoria(Categoria categoria) throws Exception;
+	public long voglioSommaPrezziDiArticoliInQuellaCategoria(Categoria categoria) throws Exception;
 	
 }
