+++ src/main/java/it/prova/gestioneordini/service/ArticoloService.java
@@ -24,6 +24,6 @@ public interface ArticoloService {
 
 	public void aggiungiCategoria(Categoria c,Articolo a) throws Exception;
 	
-	public int voglioSommaPrezziDiArticoliInQuellaCategoria(Categoria categoria) throws Exception;
+	public long voglioSommaPrezziDiArticoliInQuellaCategoria(Categoria categoria) throws Exception;
 	
 }
